NATIVE-386: Extract footers
