+++ src/routes/intro/IntroContainer.js
@@ -15,7 +15,7 @@ import styled, { type StyledComponent } from 'styled-components/native'
 import AppSettings from '../../modules/settings/AppSettings'
 import SlideContent, { type SlideContentType } from './SlideContent'
 import SentryIntegration from '../../modules/app/SentryIntegration'
-import SlideFooter from './SlideFooter'
+import SlideFooter from './footer/SlideFooter'
 import type { ViewToken } from 'react-native/Libraries/Lists/ViewabilityHelper'
 import Geolocation from '@react-native-community/geolocation'
 import CustomizableIntroSettings from './CustomizableIntroSettings'
