+++ src/routes/intro/footer/SlideButton.js
@@ -2,7 +2,7 @@
 
 import React from 'react'
 import styled, { type StyledComponent } from 'styled-components/native'
-import type { ThemeType } from '../../modules/theme/constants/theme'
+import type { ThemeType } from '../../../modules/theme/constants/theme'
 
 const ButtonContainer: StyledComponent<{}, ThemeType, *> = styled.TouchableOpacity`
   flex: 1;
