+++ src/routes/intro/footer/SlideFooter.js
@@ -1,0 +1,43 @@
+// @flow
+
+import * as React from 'react'
+import type { ThemeType } from '../../../modules/theme/constants/theme'
+import { type TFunction } from 'react-i18next'
+import type { IntroSettingsType } from '../IntroContainer'
+import StandardFooter from './StandardFooter'
+import CustomizableSettingsFooter from './CustomizableSettingsFooter'
+import SettingsFooter from './SettingsFooter'
+
+type PropsType = {|
+  slideCount: number,
+  currentSlide: number,
+  goToSlide: (index: number) => void,
+  onDone: ($Shape<IntroSettingsType>) => Promise<void>,
+  toggleCustomizeSettings: () => void,
+  customizableSettings: boolean,
+  theme: ThemeType,
+  t: TFunction
+|}
+
+class SlideFooter extends React.Component<PropsType> {
+  render () {
+    const {
+      theme, t, toggleCustomizeSettings, onDone, goToSlide, customizableSettings, currentSlide, slideCount
+    } = this.props
+
+    if (currentSlide < slideCount - 1) {
+      return <StandardFooter slideCount={slideCount} currentSlide={currentSlide} goToSlide={goToSlide} theme={theme}
+                             t={t} />
+    } else if (customizableSettings) {
+      return <CustomizableSettingsFooter slideCount={slideCount} currentSlide={currentSlide} onDone={onDone}
+                                         toggleCustomizeSettings={toggleCustomizeSettings} goToSlide={goToSlide}
+                                         theme={theme} t={t} />
+    } else {
+      return <SettingsFooter slideCount={slideCount} currentSlide={currentSlide} onDone={onDone}
+                             toggleCustomizeSettings={toggleCustomizeSettings} goToSlide={goToSlide}
+                             theme={theme} t={t} />
+    }
+  }
+}
+
+export default SlideFooter
